src/components/donor/registration/__tests__/PreferencesStep.test.tsx
```typescript
import { render, screen, fireEvent, waitFor } from '@testing-library/react';
import { vi } from 'vitest';
import PreferencesStep from '../PreferencesStep';
import { useDonorForm } from '../../../../store/donorForm';
import { useFormAnalytics } from '../../../../hooks/useFormAnalytics';

vi.mock('../../../../store/donorForm', () => ({
  useDonorForm: vi.fn()
}));

vi.mock('../../../../hooks/useFormAnalytics', () => ({
  useFormAnalytics: vi.fn()
}));

describe('PreferencesStep', () => {
  const mockUpdateFormData = vi.fn();
  const mockSetStep = vi.fn();
  const mockTrackError = vi.fn();

  beforeEach(() => {
    (useDonorForm as any).mockReturnValue({
      formData: {
        preferences: {
          newsletter: false,
          taxReceipts: false
        }
      },
      updateFormData: mockUpdateFormData,
      setStep: mockSetStep
    });

    (useFormAnalytics as any).mockReturnValue({
      trackError: mockTrackError
    });
  });

  it('renders all preference options', () => {
    render(<PreferencesStep />);

    expect(screen.getByText(/areas of interest/i)).toBeInTheDocument();
    expect(screen.getByText(/preferred communication method/i)).toBeInTheDocument();
    expect(screen.getByText(/subscribe to newsletter/i)).toBeInTheDocument();
    expect(screen.getByText(/receive tax receipts/i)).toBeInTheDocument();
  });

  it('displays validation errors', async () => {
    render(<PreferencesStep />);
    
    fireEvent.click(screen.getByText(/next step/i));

    await waitFor(() => {
      expect(screen.getByText(/select at least one interest/i)).toBeInTheDocument();
      expect(mockTrackError).toHaveBeenCalled();
    });
  });

  it('handles interest selection', async () => {
    render(<PreferencesStep />);

    const educationCheckbox = screen.getByLabelText(/education programs/i);
    fireEvent.click(educationCheckbox);

    expect(educationCheckbox).toBeChecked();
  });

  it('handles communication preference selection', () => {
    render(<PreferencesStep />);

    const select = screen.getByLabelText(/preferred communication method/i);
    fireEvent.change(select, { target: { value: 'email' } });

    expect(select).toHaveValue('email');
  });

  it('advances to next step on valid submission', async () => {
    render(<PreferencesStep />);

    // Select interests
    fireEvent.click(screen.getByLabelText(/education programs/i));
    
    // Set communication preference
    fireEvent.change(screen.getByLabelText(/preferred communication method/i), {
      target: { value: 'email' }
    });

    // Submit form
    fireEvent.click(screen.getByText(/next step/i));

    await waitFor(() => {
      expect(mockUpdateFormData).toHaveBeenCalled();
      expect(mockSetStep).toHaveBeenCalledWith(3);
    });
  });

  it('handles back button click', () => {
    render(<PreferencesStep />);
    
    fireEvent.click(screen.getByText(/back/i));
    expect(mockSetStep).toHaveBeenCalledWith(1);
  });

  it('applies correct ARIA attributes', () => {
    render(<PreferencesStep />);

    const checkboxes = screen.getAllByRole('checkbox');
    checkboxes.forEach(checkbox => {
      expect(checkbox).toHaveAttribute('aria-describedby');
    });

    const select = screen.getByRole('combobox');
    expect(select).toHaveAttribute('aria-describedby');
  });
});
```